Added new class BTModuleTreeItem.
It can be used to build module tree views easily. The code was spread and repeated in at least three places in the frontend where it does not belong. Additionally it's now possible to group modules differently and filter them easily.

ATM Search dialog module chooser uses the new class but is not quite complete.

git-svn-id: file:///home/jotik/develop/bibletime/sf_backup/svn-13.12.2011@405 ba6c7691-6512-0410-bf18-b6da21b6447f