Implemented new bookshelf models and replaced the main window bookshelf view. NB! Bookshelf view also has some changed functionality (details below).
My sincere hope is that the new bookshelf models implemented will help us clean
up large parts of the code in the near future, as they are not as cryptic as
messing around with QTreeWidgetItems. This will lead to much simpler code which
is more managable, and will let us easily implement cooler widgets using less
lines of code, as can be seen in the new bookshelf dock widget introduced in
this commit ("Code less. Create more."). :)

The idea is that the base model, BtBookshelfModel, which is based on
QAbstractListModel, will server as a base model for all other bookshelf models.
All BtBookshelfModel does nothing but provide an abstraction for a QList of
pointers to CSwordModuleInfo.

    m_bookshelfModel = new BtBookshelfModel(this);
    m_bookshelfModel->addModules(CPointers::backend()->moduleList());

The BtBookshelfTreeModel gets its data from a BtBookshelfModel. It provides an
hierarchical model for sword modules and can be used in various types of tree
views, such as BtBookshelfView. In addition, it is able to provide checking
(e.g. checkboxes in views) for all tree items. In the near future the module
checking support will be migrated to BtBookshelfModel.

    m_bookshelfTreeModel = new BtBookshelfTreeModel(this);
    // By default, check all modules:
    m_bookshelfTreeModel->setDefaultChecked(true);
    m_bookshelfTreeModel->setSourceModel(m_bookshelfModel);

By subclassing QSortFilterProxyModel, one can easily perform sorting and
filtering on models such as BtBookshelfModel or BtBookshelfTreeModel. As such,
the BtSelectedModulesBookshelfProxyModel may be used to filter out checked or
unchecked items in BtBookshelfTreeModel.

    m_filterProxyModel = new BtSelectedModulesBookshelfProxyModel(this);
    m_filterProxyModel->setSourceModel(m_bookshelfTreeModel);

A custom QTreeView subclass, BtBookshelfView, was implemented for displaying
module data in a BtBookshelfTreeModel model. It provides Qt signals for use
with popup menus etc.

    m_view = new BtBookshelfView(this);
    m_view->setModel(m_filterProxyModel);

I also changed the bookshelf view in the main window, by implementing a custom
dock widget, BtBookshelfDockWidget. Functionality-wise, the popup menus are a
bit smarter, depending whether you right-click a group or a module itself.
Since the popup menu items for changing the grouping and hiding modules are not
visible when right-clicking modules (which might be a problem, if there's
nowhere else to click) I added 2 toolbuttons for that. Showing and hiding
modules is done in-place, using checkboxes. Currently I also dropped support
for single-clicking items, because single clicking doesn't work well with the
in-place module hiding. BtBookshelfDockWidget also provides a number of quite
straightforward Qt signals for notifying the listener when opening, editing
etc operations on some module were requested by the user.


git-svn-id: file:///home/jotik/develop/bibletime/sf_backup/svn-13.12.2011@1466 ba6c7691-6512-0410-bf18-b6da21b6447f